requirements/dev.txt
# Install all of our development dependencies
-r common.txt
django-debug-toolbar==0.9.4
